requirements.txt
azure-functions
requests
beautifulsoup4==4.5.1